symgrep: Make symbol-group handling generic
The script used a different variable for each of the symbol groups, and
had a separate invocation line of the matching function for each group.
That was not very scalable so this commit changes it:

Added a symbol-group management "API" to streamline the creation of such
groups, and this is done using an associative array which holds
group:symbols pair, alongside a separate symbol group format associative
array whose keys match the former array. Both are created on group
initialization and filled in run time.

This allowed an easier iteration on the groups, less boilerplate and
less room for mistakes.